Stop heads from displaying when there is no donation associated with them

+++ src/main/java/com/enderpay/gui/DonatorsGui.java
@@ -1,12 +1,9 @@
 package com.enderpay.gui;
 
 import com.cryptomorin.xseries.SkullUtils;
-import com.cryptomorin.xseries.XBlock;
 import com.cryptomorin.xseries.XMaterial;
 import com.cryptomorin.xseries.XSound;
 import com.enderpay.Enderpay;
-import com.enderpay.model.Category;
-import com.enderpay.utils.UuidConverter;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
@@ -17,7 +14,6 @@ import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.event.inventory.InventoryDragEvent;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
-import org.bukkit.inventory.meta.SkullMeta;
 
 public class DonatorsGui extends BaseGui implements Listener {
 
@@ -42,36 +38,57 @@ public class DonatorsGui extends BaseGui implements Listener {
             ItemStack itemStack = XMaterial.PLAYER_HEAD.parseItem();
             final ItemMeta meta = itemStack.getItemMeta();
 
+            boolean isNonExistant = true;
+
             switch (i) {
                 case 0:
-                    SkullUtils.applySkin(itemStack.getItemMeta(), Enderpay.getFirstPlaceDonatorUuid());
 
-                    meta.setDisplayName(
-                            ChatColor.WHITE + Enderpay.getFirstPlaceDonatorUsername() + ChatColor.GRAY + " (" +
-                                    ChatColor.LIGHT_PURPLE + Enderpay.getCurrency().getSymbol() + Enderpay.getFirstPlaceDonatorAmount() + "" + Enderpay.getCurrency().getIso4217() + ChatColor.GRAY + ")");
+                    // check that there is a donator and that it is not N/A
+                    if (!Enderpay.getFirstPlaceDonatorUsername().equals(Enderpay.DEFAULT_PLACEHOLDER)) {
 
-                    break;
+                        SkullUtils.applySkin(itemStack.getItemMeta(), Enderpay.getFirstPlaceDonatorUuid());
+
+                        meta.setDisplayName(
+                                ChatColor.WHITE + Enderpay.getFirstPlaceDonatorUsername() + ChatColor.GRAY + " (" +
+                                        ChatColor.LIGHT_PURPLE + Enderpay.getCurrency().getSymbol() + Enderpay.getFirstPlaceDonatorAmount() + " " + Enderpay.getCurrency().getIso4217() + ChatColor.GRAY + ")");
+
+                        isNonExistant = false;
+                        break;
+                    }
                 case 1:
-                    SkullUtils.applySkin(itemStack.getItemMeta(), Enderpay.getSecondPlaceDonatorUuid());
 
-                    meta.setDisplayName(
-                            ChatColor.WHITE + Enderpay.getSecondPlaceDonatorUsername() + ChatColor.GRAY + " (" +
-                                    ChatColor.LIGHT_PURPLE + Enderpay.getCurrency().getSymbol() + Enderpay.getSecondPlaceDonatorAmount() + "" + Enderpay.getCurrency().getIso4217() + ChatColor.GRAY + ")");
+                    // check that there is a donator and that it is not N/A
+                    if (!Enderpay.getSecondPlaceDonatorUsername().equals(Enderpay.DEFAULT_PLACEHOLDER)) {
 
-                    break;
+                        SkullUtils.applySkin(itemStack.getItemMeta(), Enderpay.getSecondPlaceDonatorUuid());
+
+                        meta.setDisplayName(
+                                ChatColor.WHITE + Enderpay.getSecondPlaceDonatorUsername() + ChatColor.GRAY + " (" +
+                                        ChatColor.LIGHT_PURPLE + Enderpay.getCurrency().getSymbol() + Enderpay.getSecondPlaceDonatorAmount() + " " + Enderpay.getCurrency().getIso4217() + ChatColor.GRAY + ")");
+
+                        isNonExistant = false;
+                        break;
+                    }
                 case 2:
-                    SkullUtils.applySkin(itemStack.getItemMeta(), Enderpay.getThirdPlaceDonatorUuid());
 
-                    meta.setDisplayName(
-                            ChatColor.WHITE + Enderpay.getThirdPlaceDonatorUsername() + ChatColor.GRAY + " (" +
-                                    ChatColor.LIGHT_PURPLE + Enderpay.getCurrency().getSymbol() + Enderpay.getThirdPlaceDonatorAmount() + "" + Enderpay.getCurrency().getIso4217() + ChatColor.GRAY + ")");
+                    // check that there is a donator and that it is not N/A
+                    if (!Enderpay.getThirdPlaceDonatorUsername().equals(Enderpay.DEFAULT_PLACEHOLDER)) {
 
-                    break;
+                        SkullUtils.applySkin(itemStack.getItemMeta(), Enderpay.getThirdPlaceDonatorUuid());
+
+                        meta.setDisplayName(
+                                ChatColor.WHITE + Enderpay.getThirdPlaceDonatorUsername() + ChatColor.GRAY + " (" +
+                                        ChatColor.LIGHT_PURPLE + Enderpay.getCurrency().getSymbol() + Enderpay.getThirdPlaceDonatorAmount() + " " + Enderpay.getCurrency().getIso4217() + ChatColor.GRAY + ")");
+
+                        isNonExistant = false;
+                        break;
+                    }
             }
 
-            itemStack.setItemMeta(meta);
-
-            inventory.setItem(i, itemStack);
+            if (!isNonExistant) {
+                itemStack.setItemMeta(meta);
+                inventory.setItem(i, itemStack);
+            }
         }
 
         // add back menu item to the GUI
